requirements.txt
pillow
numpy
lib3mf
qrcode